requirements/base.txt
# Python 3.6.3
amqp==5.0.5
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
decorator==4.4.2
Django==3.1.6
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-nested-admin==3.3.3
django-redis==4.12.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
fcm-django==0.3.7
hashids==1.3.1
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
invoke==1.5.0
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
kombu==5.0.2
lxml==4.6.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
openpyxl==3.0.6
packaging==20.9
paramiko==2.7.2
parso==0.8.1
pathlib2==2.3.5
pexpect==4.8.0
phonenumbers==8.12.17
pickleshare==0.7.5
Pillow==8.1.0
plivo==4.15.2
prompt-toolkit==3.0.14
psycopg2==2.8.6
ptyprocess==0.7.0
pycparser==2.20
pycryptodome==3.9.9
pyfcm==1.4.9
Pygments==2.7.4
PyJWT==2.0.1
PyNaCl==1.4.0
pyotp==2.6.0
pyparsing==2.4.7
python-dotenv==0.15.0
python-monkey-business==1.0.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sentry-sdk==0.19.5
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0